Fix build + switch Long to Int + fix ApiKeyHandler.kt

+++ dataland-api-key-manager/src/main/kotlin/org/dataland/datalandapikeymanager/controller/ApiKeyController.kt
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 class ApiKeyController : ApiKeyAPI {
-    override fun generateApiKey(daysValid: Long?): ResponseEntity<ApiKeyData> {
+    override fun generateApiKey(daysValid: Int?): ResponseEntity<ApiKeyData> {
         return ResponseEntity.ok(ApiKeyGenerator().getNewApiKey(daysValid))
     }
